+++ src/main/java/com/boolsazo/bankchall/domain/GoalAccountPK.java
@@ -12,6 +12,6 @@ public class GoalAccountPK implements Serializable {
 
     private static final Long serialVersionUID = 1L;
 
-    private Long accountId;
-    private Long goalId;
+    private int accountId;
+    private int goalId;
 }
