+++ src/main/java/com/boolsazo/bankchall/dto/AccountResponse.java
@@ -16,11 +16,11 @@ public class AccountResponse {
     @Setter
     public static class Result {
 
-        private long accountId;
+        private int accountId;
         private String accountNumMasked; // 마스킹 된 계좌번호
         private String bankName; // 은행
         private boolean isUsed; // 이체 중인지 여부
-        private long type; // 0: 출금 1: 저축
+        private int type; // 0: 출금 1: 저축
 
         public Result(Account account, boolean isUsed) {
             this.accountId = account.getAccountId();
