+++ src/main/java/com/boolsazo/bankchall/repository/AccountRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountRepository extends JpaRepository<Account, Long> {
+public interface AccountRepository extends JpaRepository<Account, Integer> {
 
-    List<Account> findByUserIdAndType(Long userId, long type);
+    List<Account> findByUserIdAndType(int userId, int type);
 
-    Account findByAccountId(Long AccountId);
+    Account findByAccountId(int AccountId);
 
     void delete(Account account);
 }
